backend/ndis_calculator/requirements.txt
Django==2.2
django-cors-headers==2.5.2
django-filter==2.1.0
djangorestframework==3.9.2
djangorestframework_simplejwt==4.3.0
Markdown==3.1
mysqlclient==1.4.2.post1
pytz==2018.9
sqlparse==0.3.0
coverage==4.5.3
